fix(encode): align dtype to uint8 to match browser-side model

+++ scripts/encode-resorts.ts
@@ -198,7 +198,7 @@ async function encode() {
   )
 
   const extractor = await pipeline('feature-extraction', MODEL_ID, {
-    dtype: 'fp32',
+    dtype: 'uint8',
   })
 
   log('success', 'encode', 'Model loaded.')
